requirements-windows.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
deprecation==2.1.0
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
icu==0.0.1
idna==3.11
joblib==1.5.2
lxml==6.0.2
multidict==6.7.0
nltk==3.9.2
packaging==25.0
# polyglot==16.7.4  # Commented out - Windows installation issue
postgrest==2.22.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-decouple==3.8
realtime==2.22.0
regex==2025.9.18
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
